requirements.txt
pystray==0.19.0
schedule==1.2.0
Pillow==10.0.0
plyer==2.1.0